+++ CONTRIBUTING.md
@@ -40,3 +40,8 @@ python -m venv .venv
 source .venv/bin/activate
 pip install -r requirements.txt
 ```
+
+### Running Flask App
+```
+flask --app app run --debug
+```
